Remove deprecated APIs to work in the future
Remove deprecated APIs relates to handling timestamp APIs.
I checked csv formatter plugin and I used the same way
to set timestamp format for this plugin. So, column_options
like csv formatter plugin can use for this plugin's configuration.
The format can use to measure the length of the column bytes.
If it is not set, then use a default format.